test(den-web): move the dashboard tile identity unit test next to its module
#4390 added evals/specs/dashboard-element-identity.test.ts, which imports
den-web source; #4398 landed minutes later with the spec boundary ratchet
whose baseline predates that file, so dev's openwork-tests-unit went red on
the combination. Unit tests belong beside the module they test: run it in
den-web's bun suite and drop it from evals/specs.
