+++ ee/apps/den-web/package.json
@@ -10,7 +10,7 @@
     "start": "next start --hostname 0.0.0.0 --port 3005",
     "lint": "next lint",
     "typecheck": "tsc --noEmit --pretty false",
-    "test": "bun test --conditions development tests/onboarding-page.test.ts tests/openwork-models-page.test.ts tests/web-page.test.ts tests/openwork-web-billing.test.ts tests/den-api-origin.test.ts tests/den-api-redirect-rules.test.ts tests/feedback-url.test.ts tests/byok-page-primitives.test.ts tests/observability-config.test.ts tests/desktop-version-options.test.ts tests/cloud-super-admins-role-hierarchy.test.ts tests/org-dashboard-sidebar-ia.test.ts tests/dashboard-mcp-app-catalog.test.ts tests/install-errors.test.ts tests/install-download-href.test.ts tests/member-dashboard-download.test.ts tests/desktop-handoff.test.ts tests/daytona-worker-url.test.ts tests/single-org-signup-ui.test.ts tests/auth-landing-contract.test.ts tests/catalog-card-surface.test.tsx tests/mcp-connection-smart-add.test.ts tests/mcp-credential-autofill.test.ts tests/org-selection-background.test.ts tests/join-org-invite-clean.test.ts tests/scim-sso-readiness.test.ts app/api/_lib/den-api-redirect.test.mjs app/api/_lib/upstream-proxy.test.mjs app/api/_lib/upstream-proxy-lifecycle.test.mjs",
+    "test": "bun test --conditions development tests/onboarding-page.test.ts tests/openwork-models-page.test.ts tests/web-page.test.ts tests/openwork-web-billing.test.ts tests/den-api-origin.test.ts tests/den-api-redirect-rules.test.ts tests/feedback-url.test.ts tests/byok-page-primitives.test.ts tests/observability-config.test.ts tests/desktop-version-options.test.ts tests/cloud-super-admins-role-hierarchy.test.ts tests/org-dashboard-sidebar-ia.test.ts tests/dashboard-mcp-app-catalog.test.ts tests/dashboard-element-identity.test.ts tests/install-errors.test.ts tests/install-download-href.test.ts tests/member-dashboard-download.test.ts tests/desktop-handoff.test.ts tests/daytona-worker-url.test.ts tests/single-org-signup-ui.test.ts tests/auth-landing-contract.test.ts tests/catalog-card-surface.test.tsx tests/mcp-connection-smart-add.test.ts tests/mcp-credential-autofill.test.ts tests/org-selection-background.test.ts tests/join-org-invite-clean.test.ts tests/scim-sso-readiness.test.ts app/api/_lib/den-api-redirect.test.mjs app/api/_lib/upstream-proxy.test.mjs app/api/_lib/upstream-proxy-lifecycle.test.mjs",
     "test:observability": "bun test tests/observability-config.test.ts app/api/_lib/upstream-proxy.test.mjs app/api/_lib/upstream-proxy-lifecycle.test.mjs"
   },
   "dependencies": {
